Moved test for alignment.reproject, remove ipynb.
